+++ custom_components/lcars_dashboard/const.py
@@ -1,2 +1,2 @@
 DOMAIN = "lcars_dashboard"
-VERSION = "4.14.0"
+VERSION = "4.14.1"